requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.0
redis[asyncio]==5.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.2.0
python-multipart==0.0.12
prometheus-client==0.21.0
opentelemetry-sdk==1.28.4
opentelemetry-instrumentation-fastapi==0.49b4
opentelemetry-instrumentation-sqlalchemy==0.49b4
opentelemetry-exporter-otlp-proto-grpc==1.28.4
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-httpx==0.35.0
anyio==4.7.0
